Set navigation based on theme only once
