+++ app/actions/navigation/index.js
@@ -92,7 +92,6 @@ export function resetToSelectServer(allowOtherServers) {
                         },
                         options: {
                             layout: {
-                                backgroundColor: theme.centerChannelBg,
                                 componentBackgroundColor: theme.centerChannelBg,
                             },
                             statusBar: {
