+++ app/screens/edit_channel/edit_channel.js
@@ -15,7 +15,6 @@ import EventEmitter from 'mattermost-redux/utils/event_emitter';
 
 import EditChannelInfo from 'app/components/edit_channel_info';
 import {NavigationTypes, ViewTypes} from 'app/constants';
-import {setNavigatorStyles} from 'app/utils/theme';
 import {cleanUpUrlable} from 'app/utils/url';
 import {t} from 'app/utils/i18n';
 import {popTopScreen, setButtons} from 'app/actions/navigation';
@@ -65,7 +64,6 @@ export default class EditChannel extends PureComponent {
         channel: PropTypes.object.isRequired,
         currentTeamUrl: PropTypes.string.isRequired,
         updateChannelRequest: PropTypes.object.isRequired,
-        closeButton: PropTypes.object,
     };
 
     static contextTypes = {
@@ -141,10 +139,6 @@ export default class EditChannel extends PureComponent {
     }
 
     componentDidUpdate(prevProps) {
-        if (prevProps.theme !== this.props.theme) {
-            setNavigatorStyles(prevProps.componentId, this.props.theme);
-        }
-
         if (prevProps.updateChannelRequest !== this.props.updateChannelRequest) {
             switch (this.props.updateChannelRequest.status) {
             case RequestStatus.STARTED:
