+++ app/utils/theme.js
@@ -8,6 +8,21 @@ import * as ThemeUtils from 'mattermost-redux/utils/theme_utils';
 
 import {mergeNavigationOptions} from 'app/actions/navigation';
 
+const MODAL_SCREENS_WITHOUT_BACK = [
+    'AddReaction',
+    'ChannelInfo',
+    'ClientUpgrade',
+    'CreateChannel',
+    'EditPost',
+    'ErrorTeamsList',
+    'MoreChannels',
+    'MoreDirectMessages',
+    'SelectTeam',
+    'Settings',
+    'TermsOfService',
+    'UserProfile',
+];
+
 export function makeStyleSheetFromTheme(getStyleFromTheme) {
     return ThemeUtils.makeStyleFromTheme((theme) => {
         return StyleSheet.create(getStyleFromTheme(theme));
@@ -33,14 +48,17 @@ export function setNavigatorStyles(componentId, theme) {
             },
             leftButtonColor: theme.sidebarHeaderTextColor,
             rightButtonColor: theme.sidebarHeaderTextColor,
-            backButton: {
-                color: theme.sidebarHeaderTextColor,
-            },
         },
         layout: {
-            backgroundColor: theme.centerChannelBg,
+            componentBackgroundColor: theme.centerChannelBg,
         },
     };
+
+    if (!MODAL_SCREENS_WITHOUT_BACK.includes(componentId)) {
+        options.topBar.backButton = {
+            color: theme.sidebarHeaderTextColor,
+        };
+    }
 
     mergeNavigationOptions(componentId, options);
 }
